# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current auth user
      responses:
        "200":
          description: Current user or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # Dashboard summary (wow endpoint)
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects (public)
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/slug/{slug}:
    get:
      operationId: getProjectBySlug
      tags: [projects]
      summary: Get project by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

  # Skills
...
  /education:
    get:
      operationId: listEducation
      tags: [education]
      summary: List education entries (public)
      responses:
        "200":
          description: List of education
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Education"
    post:
      operationId: createEducation
      tags: [education]
      summary: Create education entry (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationInput"
      responses:
        "201":
          description: Created education
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
  /education/{id}:
    patch:
      operationId: updateEducation
      tags: [education]
      summary: Update education entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EducationUpdate"
      responses:
        "200":
          description: Updated education
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Education"
    delete:
      operationId: deleteEducation
      tags: [education]
      summary: Delete education entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Certificates
...
components:
  schemas:
...
    AuthUser:
      type: object
      required: [id, username]
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImage:
          type: ["string", "null"]
...
    Project:
      type: object
      required: [id, title, slug, shortDescription, status, category, featured, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
        description:
          type: ["string", "null"]
        problemStatement:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        features:
          type: ["string", "null"]
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: ["string", "null"]
        liveDemoUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        category:
          type: string
        status:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        screenshotUrls:
          type: array
          items:
            type: string
        demoVideoUrl:
          type: ["string", "null"]
        challenges:
          type: ["string", "null"]
        futureScope:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProjectInput:
      type: object
      required: [title, slug, shortDescription, category]
      properties:
        title:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        shortDescription:
          type: string
        description:
          type: string
        problemStatement:
          type: string
        solution:
          type: string
        features:
          type: string
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        liveDemoUrl:
          type: string
        featured:
          type: boolean
        category:
          type: string
        status:
          type: string
        thumbnailUrl:
          type: string
        screenshotUrls:
          type: array
          items:
            type: string
        demoVideoUrl:
          type: string
        challenges:
          type: string
        futureScope:
          type: string
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        problemStatement:
          type: string
        solution:
          type: string
        features:
          type: string
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        liveDemoUrl:
          type: string
        featured:
          type: boolean
        category:
          type: string
        status:
          type: string
        thumbnailUrl:
          type: string
        screenshotUrls:
          type: array
          items:
            type: string
        demoVideoUrl:
          type: string
        challenges:
          type: string
        futureScope:
          type: string
...
    Education:
      type: object
      required: [id, degree, institution, startDate, createdAt]
      properties:
        id:
          type: integer
        degree:
          type: string
        institution:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        current:
          type: boolean
        createdAt:
          type: string
    EducationInput:
      type: object
      required: [degree, institution, startDate]
      properties:
        degree:
          type: string
          minLength: 1
        institution:
          type: string
          minLength: 1
        startDate:
          type: string
        endDate:
          type: string
        grade:
          type: string
        description:
          type: string
        current:
          type: boolean
    EducationUpdate:
      type: object
      properties:
        degree:
          type: string
        institution:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        grade:
          type: string
        description:
          type: string
        current:
          type: boolean
...