# backend/src/validators/generated/types/educationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface EducationInput {
  /** @minLength 1 */
  degree: string;
  /** @minLength 1 */
  institution: string;
  startDate: string;
  endDate?: string;
  grade?: string;
  description?: string;
  current?: boolean;
}