# backend/src/validators/generated/types/educationUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface EducationUpdate {
  degree?: string;
  institution?: string;
  startDate?: string;
  endDate?: string;
  grade?: string;
  description?: string;
  current?: boolean;
}