# backend/src/validators/generated/types/projectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Portfolio CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  shortDescription: string;
  description?: string;
  problemStatement?: string;
  solution?: string;
  features?: string;
  techStack?: string[];
  githubUrl?: string;
  liveDemoUrl?: string;
  featured?: boolean;
  category: string;
  status?: string;
  thumbnailUrl?: string;
  screenshotUrls?: string[];
  demoVideoUrl?: string;
  challenges?: string;
  futureScope?: string;
}